refactor(front): Rename Features components
