+++ ui/components/imageWorkspace/src/components/Features/AutoCompleteFeatureInput.svelte
