+++ ui/components/imageWorkspace/src/components/Features/SelectFeatureInput.svelte
